+++ HISTORY.md
@@ -1,7 +1,7 @@
 unreleased
 ==========
 
-  * deps: handlebars@4.7.3
+  * deps: handlebars@4.7.6
 
 4.1.0 / 2020-01-14
 ==================